update:更改新增节点的方式

+++ slaveagent/src/main/java/com/example/agent/service/ServerDetailService.java
@@ -20,7 +20,7 @@ public class ServerDetailService {
     public static String getOS(){
         return osmxb.getName();
     }
-    public static double getMemory(){
+    public static int getMemory(){
 
         double totalvirtualMemory = osmxb.getTotalMemorySize();
         double freePhysicalMemorySize = osmxb.getFreeMemorySize();
